+++ LiteXLoader/src/Kernel/SymbolHelper.h
@@ -126,6 +126,7 @@ extern Scoreboard* g_scoreboard;
 
 //辅助函数
 BlockSource* Raw_GetBlockSourceByActor(Actor* actor);
+BlockSource* Raw_GetBlockSourceByDim(int dimid);
 int Raw_GetBlockDimension(BlockSource* bs);
 Block* Raw_GetBlockByPos(IntVec4* pos);
 Block* Raw_GetBlockByPos(int x, int y, int z, BlockSource* bs);